change approach to not have post-process-all
